requirements.txt
python-pptx==0.6.21
openai==1.3.8
pydantic==2.5.0
click==8.1.7
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
